LadderDGA.jl

LadderDGA.EnvironmentVars

LadderDGA.ImpurityQuantities

LadderDGA.ModelParameters

LadderDGA.NonLocalQuantities

LadderDGA.SimulationParameters

LadderDGA.G

LadderDGA.build_fνmax_fast

LadderDGA.calc_E_pot

LadderDGA.calc_χ_trilex

LadderDGA.expand_kGrid

LadderDGA.extend_Σ

LadderDGA.extend_γ

LadderDGA.find_usable_interval

LadderDGA.find_usable_γ

LadderDGA.gen_kGrid

LadderDGA.get_sum_helper

LadderDGA.padlength

LadderDGA.printr_s

LadderDGA.readFortranχDMFT

LadderDGA.reduce_kGrid

LadderDGA.squareLattice_ekGrid

LadderDGA.sum_freq

LadderDGA.sum_q

LadderDGA.EnvironmentVars — Type

EnvironmentVars <: ConfigStruct

Contains various settings, controlling the I/O behaviour of this module. This is typically generated from a config.toml file using the readConfig function.

LadderDGA.ImpurityQuantities — Type

ImpurityQuantities

Contains all quantities of a given channel, computed by DMFT

LadderDGA.ModelParameters — Type

ModelParameters <: ConfigStruct

Contains model parameters for the Hubbard model. This is typically generated from a config.toml file using the readConfig function.

LadderDGA.NonLocalQuantities — Type

NonLocalQuantities

Contains all non local quantities computed by the lDGA code

LadderDGA.SimulationParameters — Type

SimulationParameters <: ConfigStruct

Contains simulation parameters for the ladder DGA computations. This is typically generated from a config.toml file using the readConfig function.

Fields

TODO: describe implications of all fields

LadderDGA.G — Method

G(ind::Int64, Σ::Array{Complex{Float64},1}, ϵkGrid, β::Float64, μ)

Constructs GF from k-independent self energy, using the Dyson equation and the dispersion relation of the lattice.

LadderDGA.build_fνmax_fast — Method

build_fνmax_fast(f::AbstractArray{T,1}, nmin::Int)::Array{T, 1} where T <: Number
build_fνmax_fast(f::AbstractArray{T,2}, nmin::Int)::Array{T, 1} where T <: Number

Description

Constructs array of partial sums of one or two-dimensional array f starting at with rmin summands. This assumes, that the array is symmetric around the mid index.

Examples

julia> LadderDGA.build_fνmax_fast([3,2,1,2,3],2)
[5, 11]
julia> arr = [3 3 3 3 3; 3 2 2 2 3; 3 2 1 2 3; 3 2 2 2 3; 3 3 3 3 3
julia> LadderDGA.build_fνmax_fast(arr,2)
[17, 65]

LadderDGA.calc_E_pot — Method

Specialized function for DGA potential energy. Better performance than calc_E.

LadderDGA.calc_χ_trilex — Method

Solve χ = χ₀ - 1/β² χ₀ Γ χ ⇔ (1 + 1/β² χ₀ Γ) χ = χ₀ ⇔ (χ⁻¹ - χ₀⁻¹) = 1/β² Γ with indices: χ[ω, q] = χ₀[]

LadderDGA.expand_kGrid — Method

expand_kGrid(reducedInd, reducedArr)

Expands arbitrary Array on reduced k-Grid back to full grid. This includes restoration of mirror symmetry if the index array reducedInd indicates uncomplete grid (by having no (1,1,1) index). #Examples

    mapslices(x->expand_kGrid(qIndices, x, simParams.Nk),sdata(bubble), dims=[2])

LadderDGA.extend_Σ — Method

extend_Σ(Σ_ladder, Σ_loc, range)

Builds new self energy with Σ_loc tail, extending the frequency range of Σ_ladder. If range contains negative frequencies, the results are obtaind from get_symm_f(Σ, i).

LadderDGA.extend_γ — Method

extend_γ(arr, usable_ν)

Not tested yet!

LadderDGA.find_usable_interval — Method

find_usable_interval(arr::Array{Float64,1}; sum_type::Union{Symbol,Tuple{Int,Int}}=:common, reduce_range_prct::Float64 = 0.0)

Finds a usable interval of indices for $\chi (\omega_i)$, i.e. an interval where known physical properties hold.

Description

Returns range of indices i where the following conditions hold:

arr[i] is positive

arr[i] is decreasing (for $|i-i_0|$, $i_0$ index of 0 frequency).

Arguments

arr : array, containing $\chi (\omega_i)$

sum_type : Optional. :full will always return eachindex(), any Tuple{Int,Int} will be itnerpreted as fixed limits and the range will be fixed to these.

reduce_range_prct : Optional. Will reduce the range by given percentage (usefull to reduce effect of bad tails).

LadderDGA.find_usable_γ — Method

find_usable_γ(arr)

Not tested yet!

LadderDGA.gen_kGrid — Method

gen_kGrid(Nk, D[; min = 0, max = π, include_min=true]) =

Generates an Iterator for the Cartesian product of k vectors. This can be collected to reduce into a Nk times Nk array, containing tuples of length D.

Examples

julia> gen_kGrid(2, 2; min = 0, max = 2π, include_min = false)
Base.Iterators.ProductIterator{Tuple{Array{Float64,1},Array{Float64,1}}}(([3.141592653589793, 6.283185307179586], [3.141592653589793, 6.283185307179586]))

LadderDGA.get_sum_helper — Method

get_sum_helper(range, sP::SimulationParameters)

Description

Construct helper for (improved) sums from setting in sP over a given fit range range.

LadderDGA.padlength — Method

padlength(a,b)

computes the length of zero-padding required for convolution, using fft This is the next larger or equally large number to max(a,b) TODO: does only support padding for cube like arrays (i.e. all dimension have the same size).

Examples

julia> padlength(1:5,1:14)
8
julia> padlength(1:4,1:13)
4

LadderDGA.printr_s — Method

print 4 digits of the real part of `x`

LadderDGA.readFortranχDMFT — Method

Returns χ_DMFT[ω, ν, ν']

LadderDGA.reduce_kGrid — Method

reduce_kGrid(kGrid)

Filters the grid, so that only the lower triangle remains, i.e. for any (x1, x2, ...) the condition x1 >= x2 >= x_3 ... is fulfilled.

LadderDGA.squareLattice_ekGrid — Method

squareLattice_ek_grid(kgrid)

Computes 0.5 [cos(kx) + ... + cos(kD)] and returns a grid with Nk points.

LadderDGA.sum_freq — Method

sum_freq(f, dims::AbstractArray{Int,1}, type::T, β::Float64; 
              corr::Float64=0.0) where T <: SumHelper

Description

Wrapper for estimation of initite Matsubara sums from a finite set of samples. Computes $\frac{1}{\beta}^D \sum_{n} f(i\omega_n)$. For convenience, a correction term (being the analytic sum over a previously subtracted tail) can also be provided throught the corr argument.

Arguments

f : array containing function of Matsubara frequencies

dims : dimensions over which to sum

type : SumHelper object, used for estimation of limit

β : Inverse Temperature for normalization

corr : When tail coefficients are know, the subtraction and subsequent addition of the analytic infinite sum over the tail can be added back here.

Examples

julia> arr = ones(3,5,5)
julia> sum_freq(arr, [2,3], Naive(), 1.0)
3×1×1 Array{Float64, 3}:
    [:, :, 1] =
     25.0
     25.0
     25.0

LadderDGA.sum_q — Method

sum_q(arr, qMult)

Computes normalized sum over all q-Points.