LadderDGA.jl Documentation

Frequency Summations

K-Grids

LadderDGA.jl Documentation

Index

List of Functions

LadderDGA

LambdaCorrection

Index

LadderDGA.EnvironmentVars

LadderDGA.ModelParameters

LadderDGA.SimulationParameters

LadderDGA.γT

LadderDGA.χT

LadderDGA.χ₀T

Base.show

Base.show

LadderDGA.EPot1

LadderDGA.F_from_χ

LadderDGA.G_from_Σ

LadderDGA.G_from_Σ

LadderDGA.G_shell_sum

LadderDGA.LambdaCorrection.bisect

LadderDGA.LambdaCorrection.correct_margins

LadderDGA.LambdaCorrection.dχ_λ

LadderDGA.LambdaCorrection.find_lin_interp_root

LadderDGA.LambdaCorrection.find_root

LadderDGA.LambdaCorrection.get_λ_min

LadderDGA.LambdaCorrection.lhs_int

LadderDGA.LambdaCorrection.newton_right

LadderDGA.LambdaCorrection.λ_seach_range

LadderDGA.LambdaCorrection.λsp_rhs

LadderDGA.LambdaCorrection.χ_λ

LadderDGA.LambdaCorrection.χ_λ!

LadderDGA.OneToIndex_to_Freq

LadderDGA.bse_inv

LadderDGA.calc_E_ED

LadderDGA.calc_bubble

LadderDGA.calc_bubble_par

LadderDGA.calc_Σ_eom_par

LadderDGA.calc_Σ_par

LadderDGA.calc_λ0

LadderDGA.calc_χγ

LadderDGA.calc_χγ_par

LadderDGA.collect_Σ

LadderDGA.collect_χ

LadderDGA.collect_χ₀

LadderDGA.filling

LadderDGA.filling_pos

LadderDGA.filter_KZ

LadderDGA.filter_MA

LadderDGA.find_usable_χ_interval

LadderDGA.gen_ν_part

LadderDGA.gen_ν_part_slices

LadderDGA.gen_νω_part

LadderDGA.gen_ω_part

LadderDGA.initialize_EoM

LadderDGA.initialize_EoM_cache!

LadderDGA.iν_array

LadderDGA.iω_array

LadderDGA.log_q0_χ_check

LadderDGA.par_partition

LadderDGA.printr_s

LadderDGA.q0_index

LadderDGA.readConfig

LadderDGA.readFortranχDMFT

LadderDGA.reduce_range

LadderDGA.setup_LDGA

LadderDGA.subtract_tail

LadderDGA.subtract_tail!

LadderDGA.update_tail!

LadderDGA.update_wcache!

LadderDGA.usable_ωindices

LadderDGA.Δ

LadderDGA.Σ_Dyson

LadderDGA.λ_from_γ

LadderDGA.ν0Index_of_ωIndex

LadderDGA.νi_health

LadderDGA.νi_νngrid_pos

LadderDGA.νnGrid

LadderDGA.χ₀Asym

LadderDGA.χ₀Asym_coeffs

LadderDGA.ω0_index

LadderDGA.ω_tail

List of Functions

LadderDGA

LadderDGA.EnvironmentVars — Type

EnvironmentVars <: ConfigStruct

Contains various settings, controlling the I/O behaviour of this module. This is typically generated from a config.toml file using the readConfig function.

Fields

inputDir : String, Directory of input files

inputVars : String, File name of .jld2 file containing input.

loglevel : String, Options: disabled, error, warn, info, debug

logfile : String, Options: STDOUT, STDERR, filename

LadderDGA.ModelParameters — Type

ModelParameters <: ConfigStruct

Contains model parameters for the Hubbard model. This is typically generated from a config.toml file using the readConfig function.

Fields

U : Float64, Hubbard U

μ : Float64, chemical potential

β : Float64, inverse temperature

n : Float64, filling

sVk : Float64, ∑_k Vₖ^2

Epot_DMFT : Float64, DMFT potential energy

Ekin_DMFT : Float64, DMFT kinetic intergy

LadderDGA.SimulationParameters — Type

SimulationParameters <: ConfigStruct

Contains simulation parameters for the ladder DGA computations. This is typically generated from a config.toml file using the readConfig function.

Fields

n_iω : Int, Number of positive bosonic frequencies (full number will be 2*n_iω+1

n_iν : Int, Number of positive fermionic frequencies (full number will be 2*n_iν

n_iν_shell : Int, Number of fermionic frequencies used for asymptotic sum improvement (χ_asym_r arrays with at least these many entries need to be provided)

shift : Bool, Flag specifying if -n_iν:n_iν-1 is shifted by -ωₙ/2 at each ωₙ slice (centering the main features)

χ_helper : struct, helper struct for asymptotic sum improvements involving the generalized susceptibility (nothing if n_iν_shell == 0), see also BSE_SC.jl.

fft_range : Int, Frequencies used for computations of type f(νₙ + ωₙ).

usable_prct_reduction : Float64, percent reduction of usable bosonic frequencies

dbg_full_eom_omega : Bool, if true overrides usable ω ranges to n_iω.

LadderDGA.γT — Type

γT <: MatsubaraFunction

Struct for the non-local triangular vertex.

Fields

data : Array{ComplexF64,3}, data

axes : Dict{Symbol,Int}, Dictionary mapping :q, :ν, :ω to the axis indices.

LadderDGA.χT — Type

χT <: MatsubaraFunction

Struct for the non-local susceptibilities.

Constructor

χT(data::Array{T, 2}; full_range=true, reduce_range_prct=0.1): if full_range is set to false, the usable range is determined via find_usable_χ_interval.

Fields

data : Array{ComplexF64,3}, data

axes : Dict{Symbol,Int}, Dictionary mapping :q, :ω to the axis indices.

λ : Float64, λ correction parameter.

usable_ω : AbstractArray{Int}, usable indices for which data is assumed to be correct. See also find_usable_interval

LadderDGA.χ₀T — Type

χ₀T <: MatsubaraFunction

Struct for the bubble term. The q, ω dependent asymptotic behavior is computed from the t1 and t2 input.

Constructor

χ₀T(data::Array{_eltype,3}, kG::KGrid, ω_grid::AbstractVector{Int}, n_iν::Int, shift::Bool, mP::ModelParameters; local_tail=false)

Set local_tail=true in case of the local bubble constructed fro mthe impurity Green's function. This is necessary in order to construct the correct asymptotics.

Fields

data : Array{ComplexF64,3}, data

asym : Array{ComplexF64,2}, [q, ω] dependent asymptotic behavior.

axes : Dict{Symbol,Int}, Dictionary mapping :q, :ν, :ω to the axis indices.

Base.show — Method

Base.show(io::IO, m::ModelParameters)

Custom output for ModelParameters

Base.show — Method

Base.show(io::IO, m::SimulationParameters)

Custom output for SimulationParameters

LadderDGA.EPot1 — Method

Specialized function for DGA potential energy. Better performance than calc_E.

LadderDGA.F_from_χ — Method

F_from_χ(χ::AbstractArray{ComplexF64,3}, G::AbstractArray{ComplexF64,1}, sP::SimulationParameters, β::Float64[; diag_term=true])

TODO: documentation

LadderDGA.G_from_Σ — Method

G_from_Σ(ind::Int64, β::Float64, μ::Float64, ϵₖ::Float64, Σ::ComplexF64)
G_from_Σ(mf::ComplexF64, μ::Float64, ϵₖ::Float64, Σ::ComplexF64)

Computes Green's function according to $[\frac{(2 n + 1)\pi i}{\beta} + \mu - \epsilon_k - \Sigma(k,i\nu_n)]^{-1}$, where $\epsilon_k$ and $\Sigma(k,i\nu_n)$ are given as single values. Convenience wrappers for full grids are provided below.

Arguments:

ind : Matsubara frequency index

mf : Matsubara frequency

β : Inverse temperature (only needs to be set, if index instead of frequency is given)

μ : Chemical potential

ϵₖ : Dispersion relation at fixed k, see below for convenience wrappers.

Σ : Self energy at fixed frequency (and potentially fixed k), see below for convenience wrappers.

LadderDGA.G_from_Σ — Method

G_from_Σ(Σ::AbstractVector{ComplexF64}, ϵkGrid::Vector{Float64}, range::AbstractVector{Int}, mP::ModelParameters; μ = mP.μ,  Σloc::AbstractArray = nothing) 
G_from_Σ!(res::Matrix{ComplexF64}, Σ::AbstractVector{ComplexF64}, ϵkGrid::Vector{Float64}, range::AbstractVector{Int}, mP::ModelParameters; μ = mP.μ,  Σloc::AbstractVector = nothing)

Computes Green's function from self energy Σ and dispersion ϵkGrid over given frequency indices range. Optionally, a different chemical potential μ can be provided. When the non-local self energy is used, one typically wants to extend the usefull range of frequencies by attaching the tail of the local self energy in the high frequency regime. This is done by providing a range larger than the array size of Σ and in addition setting Σloc (the size of Σloc must be as large or larger than range). The inplace version stores the result in res.

LadderDGA.G_shell_sum — Method

G_shell_sum(N::Int, β::Float64)::Float64

Calculate $\frac{1}{\beta} \sum_{n \in \Omega_\mathrm{shell}} \frac{1}{(i \nu_n)^2}$ N should be the index of the largest frequency + 1, NOT the total lenth of the array, i.e. 50 for indices = 0:49.

LadderDGA.OneToIndex_to_Freq — Method

OneToIndex_to_Freq(ωi::Int, νi::Int, sP::SimulationParameters)

Converts (1:N,1:N) index tuple for bosonic (ωi) and fermionic (νi) frequency to Matsubara frequency number. If the array has a ν shell (for example for tail improvements) this will also be taken into account by providing Nν_shell.

LadderDGA.bse_inv — Method

bse_inv(type::Symbol, Γr::Array{ComplexF64,3})

Kernel for calculation of susceptibility and triangular vertex. Used by calc_χγ_par.

LadderDGA.calc_E_ED — Method

calc_E_ED(ϵₖ::Vector{Float64}, Vₖ::Vector{Float64}, GImp::Vector{ComplexF64}, U, n, μ, β)
calc_E_ED(ϵₖ::Vector{Float64}, Vₖ::Vector{Float64}, GImp::Vector{ComplexF64}, mP::ModelParameters)
calc_E_ED(fname::String)

Computes kinetic and potential energies from Anderson parameters.

Returns:

(EKin, EPot): Tuple{Float64,Float64}, kinetic and potential energy.

Arguments:

fname : jld2-file containing the fields: [gImp, β, ϵₖ, Vₖ, U, nden, μ] (see below)

ϵₖ : bath levels

Vₖ : hoppend amplitudes

GImp : impurity Green's function. WARNING: the arguments are assumed to by fermionic Matsuabra indices 0:length(GImp)-1!

U : Coulomb interaction strength

n : number density

μ : chemical potential

β : inverse temperature

mP : Alternative call with model parameters as Float64. See also ModelParameters.

LadderDGA.calc_bubble — Method

calc_bubble(Gνω::GνqT, Gνω_r::GνqT, kG::KGrid, mP::ModelParameters, sP::SimulationParameters; local_tail=false)

TODO: documentation

LadderDGA.calc_bubble_par — Method

calc_bubble_par(kG::KGrid, mP::ModelParameters, sP::SimulationParameters; collect_data=true)

Calculates the bubble, based on two fourier-transformed Greens functions where the second one has to be reversed.

LadderDGA.calc_Σ_eom_par — Method

calc_Σ_eom_par(νmax::Int)

Equation of motion for self energy. See calc_Σ_par.

LadderDGA.calc_Σ_par — Method

calc_Σ_par(mP::ModelParameters, sP::SimulationParameters; νmax=sP.n_iν; collect_data=true)

Calculates self-energy on worker pool. Workers must first be initialized using initialize_EoM. #TODO: νrange must be equal to the one used during initialization. remove one.

LadderDGA.calc_λ0 — Method

calc_λ0(χ₀::χ₀T, Fr::FT, χ::χT, γ::γT, mP::ModelParameters, sP::SimulationParameters)

Correction term, TODO: documentation

LadderDGA.calc_χγ — Method

calc_χγ(type::Symbol, Γr::ΓT, χ₀::χ₀T, kG::KGrid, mP::ModelParameters, sP::SimulationParameters)

Calculates susceptibility and triangular vertex in type channel. See calc_χγ_par for parallel calculation.

This method solves the following equation: `` \chir = \chi0 - \frac{1}{\beta^2} \chi0 \Gammar \chir \ \Leftrightarrow (1 + \frac{1}{\beta^2} \chi0 \Gammar) = \chi0 \ \Leftrightarrow (\chi^{-1}r - \chi^{-1}0) = \frac{1}{\beta^2} \Gamma_r ``

LadderDGA.calc_χγ_par — Method

calc_χγ_par(type::Symbol, Γr::ΓT, kG::KGrid, mP::ModelParameters, sP::SimulationParameters; collect_data=true)

Calculate susceptibility and triangular vertex parallel on workerpool.

Set collect_data to return both quantities, or call collect_χγ at a later point. calc_χγ can be used for single core computations.

LadderDGA.collect_Σ — Method

collect_Σ(Nk::Int, νrange::AbstractVector{Int}, mP::ModelParameters, sP::SimulationParameters)

LadderDGA.collect_χ — Method

collect_χ(type::Symbol, kG::KGrid, mP::ModelParameters, sP::SimulationParameters)
collect_γ(type::Symbol, kG::KGrid, mP::ModelParameters, sP::SimulationParameters)

Collects susceptibility and triangular vertex from workers, after parallel computation (see calc_χγ_par).

LadderDGA.collect_χ₀ — Method

collect_χ₀(kG::KGrid, mP::ModelParameters, sP::SimulationParameters)

Collect non-local bubble $\chi_0^{\omega}(q)$ from workers. Values first need to be calculated using calc_bubble_par.

LadderDGA.filling — Method

filling(G::Vector{ComplexF64}, [kG::KGrid, ] β::Float64)
filling(G::Vector, U::Float64, μ::Float64, β::Float64, [shell::Float64])

Computes filling of (non-) local Green's function.

If U, μ and β are provided, asymptotic corrections are used. The shell sum can be precomputed using G_shell_sum If G is defined only over positive Matsubara frequencies filling_pos can be used.

LadderDGA.filling_pos — Method

filling_pos(G::Vector, U::Float64, μ::Float64, β::Float64[, shell::Float64])::Float64

Returns filling from G only defined over positive Matsubara frequencies. See filling for further documentation.

LadderDGA.filter_KZ — Method

filter_KZ(m::Int, k::Int, X::AbstractArray{T,1}) where T <: Number

Iterated moving average noise filter for inut data. See also filter_MA.

LadderDGA.filter_MA — Method

filter_MA(m::Int, X::AbstractArray{T,1}) where T <: Number
filter_MA!(res::AbstractArray{T,1}, m::Int, X::AbstractArray{T,1}) where T <: Number

Iterated moving average noise filter for inut data. See also filter_KZ.

LadderDGA.find_usable_χ_interval — Method

find_usable_χ_interval(χ_ω::Array{Float64,1/2}; sum_type::Union{Symbol,Tuple{Int,Int}}=:common, reduce_range_prct::Float64 = 0.1)

Determines usable range for physical susceptibilities $\chi^\omega$ or $\chi^\omega$ and additionally cut away reduce_range_prct % of the range. The unusable region is given whenever the susceptibility becomes negative, or the first derivative changes sign.

Returns:

range::AbstractVector{Float64} : Usable $\omega$ range for $\chi$

Arguments:

χ_ω : $\chi^\omega$

sum_type : Optional, default :common. Can be set to :full to enforce full range, or a ::Tuple{Int,Int} to enforce a specific interval size.

reduce_range_prct : Optional, default 0.1. After finding the usable interval it is reduced by an additional percentage given by this value.

LadderDGA.gen_ν_part — Method

genνpart(νGrid::AbstractVector{Int}, sP::SimulationParameters, N::Int)

Returns partition of fermionic frequencies grid, according to the number of workers N. The partition is given as a list (of length N) of lists with 4-Tuples (ωi, ωn, νi, νn). νi and ωi are the indices for the Matsubara frequencies νn and ωn.

LadderDGA.gen_ν_part_slices — Method

gen_ν_part_slices(data::Array{ComplexF64,3}, index_list::Vector{NTuple{4,Int}})

Rearragnes data over q, ν and ω axes, for EoM (see calc_Σ_par) given index_list, which is one element of the list of lists obtained from gen_ν_part.

Returns three arrays: - datares: Rearranged data, only containes values for ν, given in `indexlist, ω values not contained indataare set to 0. - νn_list: Has lengthsize(datares,3)`. Contained fermionic Matsubara frequency for each index. - ωnranges: Has length size(data_res,3). Containes bosonic Matsubara frequencies for each ν value.

LadderDGA.gen_νω_part — Method

gen_νω_part(sP::SimulationParameters, workerpool::AbstractWorkerPool)

Returns partition of frequency grid, according to the number of workers in wp.

LadderDGA.gen_ω_part — Method

gen_ω_part(sP::SimulationParameters, workerpool::AbstractWorkerPool)

Returns partition of bosonic frequencies grid, according to the number of workers in wp.

LadderDGA.initialize_EoM — Method

function initializeEoM(Gfftreverse, λ₀::Array{ComplexF64,3}, νGrid::AbstractVector{Int}, kG::KGrid, mP::ModelParameters, sP::SimulationParameters; forcereinit = false, χsp::χT = collectχ(:sp, kG, mP, sP), γsp::γT = collectγ(:sp, kG, mP, sP), χch::χT = collectχ(:ch, kG, mP, sP), γch::γT = collectγ(:ch, kG, mP, sP))

Worker cache initialization. Must be called before calc_Σ_par.

LadderDGA.initialize_EoM_cache! — Method

initialize_EoM_cache!()

Initializes cache to correct size. kG and sP must be set first.

LadderDGA.iν_array — Method

iν_array(β::Real, grid::AbstractArray{Int64,1})::Vector{ComplexF64}
iν_array(β::Real, size::Int)::Vector{ComplexF64}

Computes list of fermionic Matsubara frequencies. If length size is given, the grid will have indices 0:size-1. Bosonic arrays can be generated with iω_array.

Returns:

Vector of fermionic Matsubara frequencies, given either a list of indices or a length.

LadderDGA.iω_array — Method

iω_array(β::Real, grid::AbstractArray{Int64,1})::Vector{ComplexF64}
iω_array(β::Real, size::Int)::Vector{ComplexF64}

Computes list of bosonic Matsubara frequencies. If length size is given, the grid will have indices 0:size-1. Fermionic arrays can be generated with iν_array.

Returns:

Vector of bosonic Matsubara frequencies, given either a list of indices or a length.

LadderDGA.log_q0_χ_check — Method

log_q0_χ_check(kG::KGrid, sP::SimulationParameters, χ::AbstractArray{_eltype,2}, type::Symbol)

TODO: documentation

LadderDGA.par_partition — Method

par_partition(set::AbstractVector, N::Int)

Returns list of indices for partition of set into N (almost) equally large segements.

LadderDGA.printr_s — Method

printr_s(x::ComplexF64)
printr_s(x::Float64)

prints 4 digits of (the real part of) x

LadderDGA.q0_index — Method

q0_index(kG::KGrid)

Index of zero k-vector.

LadderDGA.readConfig — Method

readConfig(cfg_in::String)

Reads a config.toml file either as string or from a file and returns - workerpool - ModelParameters - SimulationParameters - EnvironmentVars - kGrid (see Dispersions.jl)

LadderDGA.readFortranχDMFT — Method

Returns χ_DMFT[ω, ν, ν']

LadderDGA.reduce_range — Method

reduce_range(range::AbstractArray, red_prct::Float64)

Returns indices for 1D array slice, reduced by red_prct % (compared to initial range). Range is symmetrically reduced fro mstart and end.

LadderDGA.setup_LDGA — Method

setup_LDGA(kGridStr::Tuple{String,Int}, mP::ModelParameters, sP::SimulationParameters, env::EnvironmentVars [; local_correction=true])

Computes all needed objects for DΓA calculations.

Returns:

NTuple{16, Any}[ ΣladderLoc, Σloc, impdensity, kGrid, gLocfft, gLocrfft, Γsp, Γch, χDMFTsp, χDMFTch, χsploc, γsploc, χchloc, γch_loc, χ₀Loc, gImp ]

LadderDGA.subtract_tail! — Method

subtract_tail!(outp::AbstractArray{T,1}, inp::AbstractArray{T,1}, c::Float64, iω::Array{ComplexF64,1}, power::Int) where T <: Number

subtract the c/(iω)^power high frequency tail from inp and store in outp. See also subtract_tail

LadderDGA.subtract_tail — Method

subtract_tail(inp::AbstractArray{T,1}, c::Float64, iω::Array{ComplexF64,1}, power::Int) where T <: Number

subtract the $\frac{c}{(i\omega)^\text{power}}$ high frequency tail from input array inp.

LadderDGA.update_tail! — Method

update_tail!(χ::χT, new_tail_c::Array{Float64}, ωnGrid::Array{ComplexF64})

Updates the $\frac{c_i}{\omega_n^i}$ tail for all coefficients given in new_tail_c (index 1 corresponds to $i=0$).

LadderDGA.update_wcache! — Method

update_wcache!(name::Symbol, val; override=true)

Updates worker cache with given name and value. Typically used through remotecall() on specific worker.

LadderDGA.usable_ωindices — Method

usable_ωindices(sP::SimulationParameters, χ_sp::χT, χ_ch::χT)

Helper function, returning the indices n for $\omega_n$ ranges of multiple channels. If dbg_full_eom_omega is set to true in the config, the full range will be returned, otherwise an intersection of the usable ranges obtained from find_usable_χ_interval.

LadderDGA.Δ — Method

Δ(ϵₖ::Vector{Float64}, Vₖ::Vector{Float64}, νₙ::Vector{ComplexF64})::Vector{ComplexF64}

Computes hybridization function $\Delta(i\nu_n) = \sum_k \frac{|V_k|^2}{\nu_n - \epsilon_k}$ from Anderson parameters (for example obtained through exact diagonalization).

Returns:

Hybridization function over list of given fermionic Matsubara frequencies.

Arguments:

ϵₖ : list of bath levels

Vₖ : list of hopping amplitudes

νₙ : Vector of fermionic Matsubara frequencies, see also: iν_array.

LadderDGA.Σ_Dyson — Method

Σ_Dyson(GBath::Vector{ComplexF64}, GImp::Vector{ComplexF64})::Vector{ComplexF64}
Σ_Dyson!(Σ::Vector{ComplexF64}, GBath::Vector{ComplexF64}, GImp::Vector{ComplexF64})::Vector{ComplexF64}

Calculates $\Sigma = 1 / G_\text{bath} - 1 / G_\text{imp}$.

LadderDGA.λ_from_γ — Method

λ_from_γ(type::Symbol, γ::γT, χ::χT, U::Float64)

TODO: documentation

LadderDGA.ν0Index_of_ωIndex — Method

ν0Index_of_ωIndex(ωi::Int[, sP])::Int

Calculates index of zero fermionic Matsubara frequency (which may depend on the bosonic frequency). ωi is the index (i.e. starting with 1) of the bosonic Matsubara frequency.

LadderDGA.νi_health — Method

νi_health(νGrid::AbstractArray{Int}, sP::SimulationParameters)

Returns a list of available bosonic frequencies for each fermionic frequency, given in νGrid. This can be used to estimate the maximum number of usefull frequencies for the equation of motion.

LadderDGA.νi_νngrid_pos — Method

νi_νngrid_pos(ωi::Int, νmax::Int, sP::SimulationParameters)

Indices for positive fermionic Matsubara frequencies, depinding on ωi, the index of the bosonic Matsubara frequency.

LadderDGA.νnGrid — Method

νnGrid(ωn::Int, sP::SimulationParameters)

Calculates grid of fermionic Matsubara frequencies for given bosonic frequency ωn (including shift, if set through sP).

LadderDGA.χ₀Asym — Method

χ₀Asym(c1::Float64, c2::Vector{Float64}, c3::Float64, ω_grid::AbstractVector{Int}, n_iν::Int, shift::Int, β::Float64)

Builds asymtotic helper array. See calc_bubble implementation for details.

c1, c2 and c3 are the coefficients for the asymtotic tail expansion and can be obtained through χ₀Asym_coeffs. n_iν is the number of positive fermionic Matsubara frequencies, shift is either 1 or 0, depending on the type of freqiency grid.

LadderDGA.χ₀Asym_coeffs — Method

χ₀Asym_coeffs(kG::KGrid, local_tail::Bool, mP::ModelParameters)

Builds tail coefficients for the χ₀ asymptotic helper, obtained through χ₀Asym.

LadderDGA.ω0_index — Method

ω0_index(sP::SimulationParameters)
ω0_index(χ::[χT or AbstractMatrix])

Index of ω₀ frequency.

LadderDGA.ω_tail — Method

ω_tail(ωindices::AbstractVector{Int}, coeffs::AbstractVector{Float64}, sP::SimulationParameters) 
ω_tail(χ_sp::χT, χ_ch::χT, coeffs::AbstractVector{Float64}, sP::SimulationParameters)

LambdaCorrection

This sub-module contains function related to the $\lambda$-correction.

LadderDGA.LambdaCorrection.bisect — Method

bisect(λl::T, λm::T, λr::T, Fm::T)::Tuple{T,T} where T <: Union{Float64, Vector{Float64}}

WARNING: Not properly tested! Bisection root finding algorithm. This is a very crude adaption of the 1D case. The root may therefore lie outside the given region and the search space has to be corrected using correct_margins.

Returns:

(Vector of) new interval borders, according to Fm.

Arguments:

λl : (Vector of) left border(s) of bisection area

λm : (Vector of) central border(s) of bisection area

λr : (Vector of) right border(s) of bisection area

Fm : (Vector of) Poincare-Miranda condition (s)

LadderDGA.LambdaCorrection.correct_margins — Method

correct_margins(λl::T, λm::T, λr::T, Fm::T, Fr::T)::Tuple{T,T} where T <: Union{Float64, Vector{Float64}}

Helper method for bisect.

LadderDGA.LambdaCorrection.dχ_λ — Method

dχ_λ(χ::[Float64,ComplexF64,AbstractArray], λ::Float64)

First derivative of χ_λ.

LadderDGA.LambdaCorrection.find_lin_interp_root — Method

find_lin_interp_root(xdata::AbstractVector{Float64}, ydata::AbstractVector{Float64})

WARNING: this is a specialized function which assumes strictly monotonic data! Given sampled xdata and ydata, find the root using linear interpolation. Returns estimated x₀.

Returns:

x₀ : Float64, root of sampled function data.

Arguments:

xdata : x-data of samples from strictly monotonic decreasing function.

ydata : y-data of samples from strictly monotonic decreasing function

LadderDGA.LambdaCorrection.find_root — Method

find_root(c2_data::Array{Float64,2})

Determines root (i.e. $(\lambda_\text{sp}, \lambda_\text{ch})$) from residual curve computed via residuals. The root is determined via linear interpolation from the given data.

Returns:

$(\lambda_\text{sp}, \lambda_\text{ch}, \text{check})$, roots for $\lambda$ values and a check (Boolean). Check is false, if the $\lambda$ values could not be determined (NaN )

Arguments:

c2_data : Matrix{Float64}, data from residuals.

LadderDGA.LambdaCorrection.get_λ_min — Method

get_λ_min(χr::AbstractArray{Float64,2})::Float64

Computes the smallest possible $\lambda$-correction parameter (i.e. first divergence of $\chi(q)$), given as $\lambda_\text{min} = - \min_{q}(1 / \chi^{\omega_0}_q)$.

LadderDGA.LambdaCorrection.lhs_int — Method

lhs_int(χ_sp::Matrix, χ_ch::Matrix, χ_tail::Vector{ComplexF64}, kMult::Vector{Float64}, k_norm::Int, Ekin::Float64, β::Float64)

Internal function. This calculates the sum over up-up and up-down susceptibilities, used in cond_both_int, avoiding allocations.

Returns:

Tuple{ComplexF64/Float64}[ lhsc1, # ``\sum{q,\omega} \chi{\uparrow,\downarrow}`` lhsc2 # $\sum_{q,\omega} \chi_{\uparrow,\uparrow}$ ]

LadderDGA.LambdaCorrection.newton_right — Method

newton_right(f::Function, df::Function, start::[Float64,Vector{Float64}; nsteps=5000, atol=1e-11)

WARNING: Not properly tested! This is an adaption of the traditional Newton root finding algorithm, searching only to the right of start.

LadderDGA.LambdaCorrection.λ_seach_range — Method

λ_seach_range(χ::Matrix{Float64}; λ_max_default = 50)

Calculates reasonable interval for the search of the $\lambda$-correction parameter.

The interval is chosen with $\lambda_\text{min}$, such that all unphysical poles are excluded and $\lambda_\text{max} = \lambda_\text{default} / \max_{q,\omega} \chi(q,\omega)$. The λ_max_default parameter may need to be adjusted, depending on the model, since in principal arbitrarily large maximum values are possible.

LadderDGA.LambdaCorrection.λsp_rhs — Function

λsp_rhs([imp_density::Float64, ]χsp::χT, χch::χT, kG::KGrid, mP::ModelParameters, sP::SimulationParameters, λ_rhs = :native)

Helper function for the right hand side of the Pauli principle conditions (old λ correction). TODO: write down formula, explain imp_density as compensation to DMFT.

LadderDGA.LambdaCorrection.χ_λ! — Method

χ_λ!(χ_destination::[AbstractArray,χT], [χ::[AbstractArray,χT], ] λ::Float64)

Inplace version of χ_λ. If the second argument is omitted, results are stored in the input χ.

LadderDGA.LambdaCorrection.χ_λ — Method

χ_λ(χ::[Float64,ComplexF64,AbstractArray,χT], λ::Float64)

Computes the λ-corrected susceptibility: $\chi^{\lambda,\omega}_q = \frac{1}{1 / \chi^{\lambda,\omega}_q + \lambda}$. The susceptibility $\chi$ can be either given element wise, or as χT See also χT in LadderDGA.jl.